Running thread now works
Running programs in a thread functionality is implemented.
Results are displayed in the terminal below the code, and the run menu
is only available if the compiled program is ready. Now the program
variable is never null, instead when initializing the task, the compiler
can return a precompiled program too!