Now the GUI caches everything, and every data for possible tasks and task collections are directly available in the elements.
Previously there was a disconnection between the View and the Model, and
every value for the View must been queried from the model. Now the Model
creates a cache, and updates the View only once, and then the view does
not need to poll the controller, or the user for these basic data.